# api/openapi/api.json
{"openapi":"3.0.0","info":{"title":"ひまわりプロジェクト","description":"APIドキュメント","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"TaskSerializer":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"}},"required":["id","title","description","status"]},"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"}},"required":["title","description"]},"SignUpUserDto":{"type":"object","properties":{"username":{"type":"string","example":"test1","minLength":5,"maxLength":20},"email":{"type":"string","example":"test1@gmail.com"},"password":{"type":"string","example":"hogehoge","minLength":6,"maxLength":20}},"required":["username","email","password"]},"SignInUserDto":{"type":"object","properties":{"username":{"type":"string","example":"test1","minLength":5,"maxLength":20},"email":{"type":"string","example":"test1@gmail.com"},"password":{"type":"string","example":"hogehoge","minLength":6,"maxLength":20}},"required":["password"]},"AccessTokenSerializer":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"UserSerializer":{"type":"object","properties":{"id":{"type":"number"},"username":{"type":"string"},"email":{"type":"string"},"avatarUrl":{"type":"string"},"statusMessage":{"type":"string"},"groups":{"type":"array","items":{"$ref":"#/components/schemas/GroupSerializer"}}},"required":["id","username","email"]},"CommitSerializer":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"studyHours":{"type":"number"},"studyMinutes":{"type":"number"},"goalId":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"goal":{"$ref":"#/components/schemas/GoalSerializer"}},"required":["id","title","description","studyHours","studyMinutes","goalId","createdAt"]},"GoalSerializer":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string"},"label":{"type":"string","enum":["CHALLENGING","ACHIEVEMENT","GIVE_UP"]},"isPublic":{"type":"boolean"},"userId":{"type":"number"},"lastCommitedAt":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"},"totalTime":{"type":"string"},"user":{"$ref":"#/components/schemas/UserSerializer"},"commits":{"type":"array","items":{"$ref":"#/components/schemas/CommitSerializer"}}},"required":["id","title","description","label","isPublic","userId","lastCommitedAt","createdAt"]},"GroupSerializer":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"users":{"type":"array","items":{"$ref":"#/components/schemas/UserSerializer"}},"goals":{"type":"array","items":{"$ref":"#/components/schemas/GoalSerializer"}},"lastTimelinePostedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","lastTimelinePostedAt"]},"UpdateMeDto":{"type":"object","properties":{"username":{"type":"string","example":"updated_test1","minLength":5,"maxLength":20},"avatarUrl":{"type":"string","example":"http://i.pravatar.cc/64"},"statusMessage":{"type":"string","example":"休暇中です","minLength":0,"maxLength":200}},"required":["username"]},"UserAndTokenSerializer":{"type":"object","properties":{"me":{"$ref":"#/components/schemas/UserSerializer"},"accessToken":{"type":"string"}},"required":["me","accessToken"]},"ResendVerifyEmailDto":{"type":"object","properties":{"email":{"type":"string","example":"test1@gmail.com"}},"required":["email"]},"CreateGoalDto":{"type":"object","properties":{"title":{"type":"string","example":"TOIEC800点以上取る"},"description":{"type":"string","example":"2021年度の初回実施日に取りたい"},"isPublic":{"type":"boolean","description":"publicに公開するか否か","default":false}},"required":["title","isPublic"]},"UpdateGoalDto":{"type":"object","properties":{"title":{"type":"string","example":"TOIEC800点以上取る"},"description":{"type":"string","example":"2021年度の初回実施日に取りたい"},"isPublic":{"type":"boolean","description":"publicに公開するか否か","default":false},"label":{"type":"string","description":"ラベル","example":"CHALLENGING","enum":["CHALLENGING","ACHIEVEMENT","GIVE_UP"]}},"required":["title","isPublic","label"]},"CreateCommitDto":{"type":"object","properties":{"title":{"type":"string","example":"Writingのテスト問題集","minLength":1,"maxLength":20},"description":{"type":"string","example":""},"studyHours":{"type":"number","example":5,"maximum":23,"minimum":0},"studyMinutes":{"type":"number","example":16,"maximum":59,"minimum":0}},"required":["title","studyHours","studyMinutes"]},"MonthlyCount":{"type":"object","properties":{"createdAt":{"type":"string","example":"2020-06"},"count":{"type":"number","example":3}},"required":["createdAt","count"]},"CommitsSummary":{"type":"object","properties":{"totalTime":{"type":"string","example":"88:24"},"totalCount":{"type":"number","example":99}},"required":["totalTime","totalCount"]},"CreateGroupDto":{"type":"object","properties":{"name":{"type":"string","example":"グループひまわり","minLength":1,"maxLength":20},"emails":{"example":["test2@gmail.com","test3@gmail.com"],"type":"array","items":{"type":"string"}}},"required":["name","emails"]},"InviteUserDto":{"type":"object","properties":{"email":{"type":"string","example":"test2@gmail.com"}},"required":["email"]},"InviteUsersDto":{"type":"object","properties":{"emails":{"example":["test2@gmail.com","test3@gmail.com"],"type":"array","items":{"type":"string"}}},"required":["emails"]},"AssignGoalDto":{"type":"object","properties":{"goalId":{"type":"number","example":1}},"required":["goalId"]},"BulkAssignGoalsDto":{"type":"object","properties":{"goalIds":{"description":"紐づけるゴールのID","example":[1,2,4],"type":"array","items":{"type":"number"}}},"required":["goalIds"]},"CommentSerializer":{"type":"object","properties":{"id":{"type":"number"},"content":{"type":"string"},"userId":{"type":"number"},"timelineId":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/UserSerializer"},"timeline":{"$ref":"#/components/schemas/TimelineSerializer"}},"required":["id","content","userId","timelineId","createdAt"]},"TimelineSerializer":{"type":"object","properties":{"id":{"type":"number"},"type":{"type":"string","enum":["COMMIT_CREATED","GOAL_STATUS_UPDATED"]},"commit":{"$ref":"#/components/schemas/CommitSerializer"},"goal":{"$ref":"#/components/schemas/GoalSerializer"},"fromLabel":{"type":"string","enum":["CHALLENGING","ACHIEVEMENT","GIVE_UP"]},"toLabel":{"type":"string","enum":["CHALLENGING","ACHIEVEMENT","GIVE_UP"]},"reactions":{"type":"array","items":{"$ref":"#/components/schemas/ReactionSerializer"}},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentSerializer"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","type","fromLabel","toLabel","createdAt","updatedAt"]},"ReactionSerializer":{"type":"object","properties":{"id":{"type":"number"},"emoji":{"type":"string","enum":["GOOD","BAD","SMILE","PIEN","POPPER"]},"timelineId":{"type":"number"},"userId":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"timeline":{"$ref":"#/components/schemas/TimelineSerializer"},"user":{"$ref":"#/components/schemas/UserSerializer"}},"required":["id","emoji","timelineId","userId","createdAt"]},"CreateCommentDto":{"type":"object","properties":{"content":{"type":"string","example":"ファイト！"}},"required":["content"]},"CreateReactionDto":{"type":"object","properties":{"emoji":{"type":"string","example":"PIEN","enum":["GOOD","BAD","SMILE","PIEN","POPPER"]}},"required":["emoji"]},"SearchSerializer":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/UserSerializer"}},"goals":{"type":"array","items":{"$ref":"#/components/schemas/GoalSerializer"}}},"required":["users","goals"]}}},"paths":{"/api/tasks":{"get":{"operationId":"TasksController_getTasks","parameters":[],"responses":{"200":{"description":"タスクの一覧を取得できた","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskSerializer"}}}}}},"tags":["tasks"],"security":[{"bearer":[]}]},"post":{"operationId":"TasksController_createTask","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"タスクが正常に作成されました","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskSerializer"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]}},"/api/tasks/{id}":{"get":{"operationId":"TasksController_getTaskById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"タスクを1件取得できた","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskSerializer"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]},"delete":{"operationId":"TasksController_deleteTask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"タスクが削除された"}},"tags":["tasks"],"security":[{"bearer":[]}]}},"/api/tasks/{id}/status":{"patch":{"operationId":"TasksController_updateTaskStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"タスクのステータスが更新できた","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskSerializer"}}}}},"tags":["tasks"],"security":[{"bearer":[]}]}},"/api/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpUserDto"}}}},"responses":{"201":{"description":"ユーザー登録完了(メール送信)"},"400":{"description":"入力値のフォーマットエラー"},"409":{"description":"usernameまたはemailの重複エラー"}},"tags":["auth"]}},"/api/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInUserDto"}}}},"responses":{"200":{"description":"ユーザーログイン完了","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessTokenSerializer"}}}},"401":{"description":"メール認証ができてない、もしくはusername(email)またはpasswordが違う"}},"tags":["auth"]}},"/api/auth/google":{"get":{"operationId":"AuthController_googleLogin","parameters":[],"responses":{"200":{"description":"グーグルログイン"}},"tags":["auth"]}},"/api/auth/google/callback":{"get":{"operationId":"AuthController_googleLoginCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/email/verify/{token}":{"get":{"operationId":"AuthController_verifyEmail","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"ユーザー本登録完了"},"400":{"description":"メール認証済"},"404":{"description":"無効なトークン"}},"tags":["auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"ログインユーザー自身の情報を取得","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSerializer"}}}}},"tags":["auth"],"security":[{"bearer":[]}]},"put":{"operationId":"AuthController_updateMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMeDto"}}}},"responses":{"200":{"description":"ログインユーザー自身の情報を更新する","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAndTokenSerializer"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/api/auth/email/resend":{"post":{"operationId":"AuthController_resendEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendVerifyEmailDto"}}}},"responses":{"200":{"description":"認証メール再送信完了"},"404":{"description":"ユーザーが存在しない、またはメール認証済み、または外部サービス認証"}},"tags":["auth"]}},"/api/goals/summary/monthly":{"get":{"operationId":"GoalsController_getSummary","parameters":[],"responses":{"200":{"description":"目標と学習記録の月ごとのサマリーを取得する","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["goals"],"security":[{"bearer":[]}]}},"/api/goals":{"post":{"operationId":"GoalsController_createGoal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGoalDto"}}}},"responses":{"201":{"description":"目標の作成","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoalSerializer"}}}}},"tags":["goals"],"security":[{"bearer":[]}]},"get":{"operationId":"GoalsController_getGoals","parameters":[],"responses":{"200":{"description":"目標の一覧取得","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GoalSerializer"}}}}}},"tags":["goals"],"security":[{"bearer":[]}]}},"/api/goals/{id}":{"get":{"operationId":"GoalsController_getGoal","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"目標の詳細取得","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoalSerializer"}}}}},"tags":["goals"],"security":[{"bearer":[]}]},"put":{"operationId":"GoalsController_updateGoal","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGoalDto"}}}},"responses":{"200":{"description":"目標情報の更新","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoalSerializer"}}}}},"tags":["goals"],"security":[{"bearer":[]}]}},"/api/goals/{goal_id}/commits":{"post":{"operationId":"CommitsController_createCommit","parameters":[{"name":"goal_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommitDto"}}}},"responses":{"201":{"description":"学習記録の作成","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommitSerializer"}}}}},"tags":["commits"],"security":[{"bearer":[]}]}},"/api/commits":{"get":{"operationId":"CommitsController_getCommits","parameters":[],"responses":{"200":{"description":"ログインユーザーの全学習履歴の取得","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommitSerializer"}}}}}},"tags":["commits"],"security":[{"bearer":[]}]}},"/api/commits/summary/monthly":{"get":{"operationId":"CommitsController_getMonthlyCountByUser","parameters":[],"responses":{"200":{"description":"ログインユーザーの月単位での全学習記録数の取得","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MonthlyCount"}}}}}},"tags":["commits"],"security":[{"bearer":[]}]}},"/api/commits/summary/total":{"get":{"operationId":"CommitsController_getSummaryByUser","parameters":[],"responses":{"200":{"description":"累計学習時間等の取得","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommitsSummary"}}}}},"tags":["commits"],"security":[{"bearer":[]}]}},"/api/commits/{id}":{"delete":{"operationId":"CommitsController_deleteCommit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":"学習記録の単体削除"},"404":{"description":"投稿ユーザー以外による操作、または目標と学習記録が紐付いていない場合"}},"tags":["commits"],"security":[{"bearer":[]}]}},"/api/groups":{"post":{"operationId":"GroupsController_createGroup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupDto"}}}},"responses":{"201":{"description":"グループの作成","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupSerializer"}}}},"400":{"description":"存在しないメールアドレスが含まれていた場合"}},"tags":["groups"],"security":[{"bearer":[]}]},"get":{"operationId":"GroupsController_getGroups","parameters":[],"responses":{"200":{"description":"参加しているグループ一覧を取得","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GroupSerializer"}}}}}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}/users":{"post":{"operationId":"GroupsController_inviteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteUserDto"}}}},"responses":{"201":{"description":"グループへの招待"}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}/users/multiple":{"post":{"operationId":"GroupsController_inviteUsers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InviteUsersDto"}}}},"responses":{"201":{"description":"グループへの複数のメールアドレスでの招待"},"400":{"description":"存在しないメールアドレスが含まれていた場合"}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}":{"get":{"operationId":"GroupsController_getGroup","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"グループの基本情報を取得","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupSerializer"}}}},"404":{"description":"参加していない or 存在していないグループを指定した時404"}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}/goals":{"post":{"operationId":"GroupsController_assignGoal","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignGoalDto"}}}},"responses":{"201":{"description":"グループへの目標登録"},"409":{"description":"目標をグループに登録する際に、登録済だった場合"}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}/goals/bulk":{"put":{"operationId":"GroupsController_bulkAssignGoals","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkAssignGoalsDto"}}}},"responses":{"200":{"description":"グループへの目標の一括登録。ここに含まれていない目標IDに関しては、すでに登録済だった場合、自動で削除する","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupSerializer"}}}},"400":{"description":"他人の目標のIDを操作しようとした時"}},"tags":["groups"],"security":[{"bearer":[]}]}},"/api/groups/{id}/timelines":{"get":{"operationId":"TimelinesController_getTimelines","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"timeline取得用ダミーAPI","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TimelineSerializer"}}}}}},"tags":["timelines"],"security":[{"bearer":[]}]}},"/api/timelines/{id}/comments":{"post":{"operationId":"CommentsController_createComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentDto"}}}},"responses":{"201":{"description":"コメントの投稿","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentSerializer"}}}}},"tags":["comments"],"security":[{"bearer":[]}]}},"/api/timelines/{id}/reactions":{"post":{"operationId":"ReactionsController_createReaction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReactionDto"}}}},"responses":{"201":{"description":"リアクションの投稿","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReactionSerializer"}}}},"204":{"description":"既に投稿しているリアクションを削除"},"400":{"description":"許可されていない絵文字を入力した場合"}},"tags":["reactions"],"security":[{"bearer":[]}]}},"/api/searches/in_related_users":{"get":{"operationId":"SearchesController_searchInGroupRelatedUsers","parameters":[],"responses":{"200":{"description":"自分の参加しているグループのメンバーの中での検索結果","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchSerializer"}}}}},"tags":["searches"],"security":[{"bearer":[]}]}},"/api/searches/users":{"get":{"operationId":"SearchesController_getUsers","parameters":[{"name":"keyword","required":true,"in":"query","example":"test","schema":{"type":"string"}}],"responses":{"200":{"description":"全体検索(ユーザー)","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserSerializer"}}}}}},"tags":["searches"],"security":[{"bearer":[]}]}},"/api/searches/goals":{"get":{"operationId":"SearchesController_getGoals","parameters":[],"responses":{"200":{"description":"全体検索(目標)","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GoalSerializer"}}}}}},"tags":["searches"],"security":[{"bearer":[]}]}},"/api/users/{id}":{"get":{"operationId":"UsersController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"他人の基本情報を取得する","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSerializer"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/{id}/goals":{"get":{"operationId":"UsersController_getGoalsOfUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"他人の公開目標一覧を取得する","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GoalSerializer"}}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/{id}/commits/summary/monthly":{"get":{"operationId":"UsersController_getMonthlyCountByUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"他人の月単位での全学習記録数の取得","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MonthlyCount"}}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/{id}/commits/summary/total":{"get":{"operationId":"UsersController_getCommitSummaryByUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"他人の累計学習時間と累計学習記録数の取得","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommitsSummary"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/users/{id}/goals/summary/monthly":{"get":{"operationId":"UsersController_getGoalCommitMonthlySummary","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"目標と学習記録の月ごとのサマリーを取得する（公開目標のみ）","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"],"security":[{"bearer":[]}]}}}}