# api/src/auth/serializer/user-and-token.serializer.ts
import { ApiProperty } from '@nestjs/swagger';
import { UserSerializer } from './user.serializer';

export class UserAndTokenSerializer {
  @ApiProperty()
  me: UserSerializer;

  @ApiProperty()
  accessToken: string;
}